Cleaned up the codebase by suppressing diagnostic warnings, streamlining null-safety assertions, and updating Kotlin/Native exception handling configuration.
*   **Compiler Warning Management:**
    *   Applied bulk `@file:Suppress` to the `omnifret-gplayer` module to silence redundant compiler diagnostics, including `REDUNDANT_CALL_OF_CONVERSION_METHOD`, `UNNECESSARY_LATEINIT`, `CAST_NEVER_SUCCEEDS`, and `SENSELESS_COMPARISON`.
    *   Added `@Suppress("DEPRECATION")` for `LocalClipboardManager` in `SettingsScreen.kt` and `LocalLifecycleOwner` in `MicPermission.kt`, including documentation on future migration paths.

*   **Build & Kotlin/Native Configuration:**
    *   Removed the deprecated `-Xforeign-exception-mode=objc-wrap` global flag from `shared/build.gradle.kts`.
    *   Updated the Kotlin/Native exception handling documentation to favor per-library cinterop modes and explicit `try/catch` blocks in entries calling into Apple frameworks (e.g., `AVAudioFile`).
    *   Added `@OptIn` for `ExperimentalUnsignedTypes` in `EnvironmentPartials.kt`.

*   **Audio Engine & Rendering Cleanup:**
    *   Streamlined `PipelinedRenderLoop`, `ChunkRenderer`, and `IosBackingTrackEngine` by removing redundant null-safety assertions (`!!`) and unnecessary null checks on non-nullable platform types.
    *   Refactored `AVAudioPCMBuffer` and `AVAudioFormat` instantiations across iOS-specific audio components to remove redundant optional handling where the API bindings are now guaranteed.
    *   Updated Skia rendering logic in `PlatformTextSupport.ios.kt` to use `canvas.skiaCanvas` instead of `canvas.nativeCanvas`.

*   **UI & Tooling Updates:**
    *   Updated `PdfViewerScreen` to match the revised `rememberTransformableState` lambda signature.
    *   Optimized type conversions in `Set.kt`, `TabNotationView.ios.kt`, and `JamScreenV2Native.ios.kt` by removing redundant `toDouble()` calls.
    *   Simplified PDF document loading in `PdfDocument.ios.kt` by removing unnecessary error branching.

+++ gplayer/src/commonMain/kotlin/com/omnifret/gplayer/core/ecmaScript/Set.kt
@@ -10,7 +10,7 @@ public class Set<T> : Iterable<T> {
     }
 
     public val size: Double
-        get() = _storage.size.toDouble()
+        get() = _storage.size
 
     public constructor(values: Iterable<T>?) {
         _storage = ObjectBooleanMap()
